// src/app/core/services/model/reportOptionsQueryOutputDto.ts
/**
 * TWMS REST API
 * TWMS REST API
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

/**
 * ouput object for the report options query (filter for device properties)
 */
export interface ReportOptionsQueryOutputDto { 
    /**
     * name of the device point
     */
    devicePointName?: string;
    /**
     * uuid of the device point
     */
    devicePointUuid?: string;
    /**
     * name of the device
     */
    deviceName?: string;
    /**
     * uuid of the device
     */
    deviceUuid?: string;
    /**
     * type of the device, e.g. \"Wall-E\"
     */
    deviceTypeName?: string;
    /**
     * number of the channel, e.g. 6 (6 = Temperature_Get, see enum ChannelValueCodes or emico c api)
     */
    dataPointId?: string;
    /**
     * human-readable channel name  datapoint description
     */
    description?: string;
    /**
     * datapoint Unit
     */
    unit?: string;
    /**
     * datapoint DataType
     */
    dataType?: string;
    /**
     * install timestamp of the device
     */
    deviceInstallTimestamp?: string;
    /**
     * uninstall timestamp of the device
     */
    deviceUninstallTimestamp?: string;
}