// src/app/core/services/model/reportOptionsQueryResultDto.ts
/**
 * TWMS REST API
 * TWMS REST API
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { ReportOptionsQueryOutputDto } from './reportOptionsQueryOutputDto';

/**
 * extended ouput object for the report options query (filter for device properties)
 */
export interface ReportOptionsQueryResultDto { 
    /**
     * options for the given query, e.g. \"DevicePointName\" = new List() {\"Beate 1\", \"Trinkwassererwärmer 7\"}
     */
    dropdownOptions?: { [key: string]: Array<string>; };
    /**
     * device info (for all devices returned by the report options query)
     */
    detailedDeviceInfo?: Array<ReportOptionsQueryOutputDto>;
}